Agregue Service de Adopcion, Visitas de Seguimiento y clase Seguimiento

+++ IntegradorPA2/src/main/java/com/mycompany/integradorpa2/igu/Gatos/ReadGato.java
@@ -90,7 +90,7 @@ public class ReadGato extends javax.swing.JFrame {
         modelo.setRowCount(0); // limpiar
         List<Gato> gatos = gatoDao.listarTodos();
         for (Gato g : gatos) {
-            Long id        = g.getId();
+            int id        = g.getId();
             String nombre  = nvl(g.getNombre());
             String raza    = nvl(g.getRaza());
             Integer edad   = g.getEdad() != null ? g.getEdad() : 0;
