+++ IntegradorPA2/src/main/java/com/mycompany/integradorpa2/logica/Gato.java
@@ -19,7 +19,7 @@ import javax.persistence.OneToOne;
 public class Gato {
     @Id
     @GeneratedValue(strategy=GenerationType.SEQUENCE)
-    private Long id;
+    private int id;
     @Basic
     private String nombre;
     private String raza;
@@ -51,7 +51,7 @@ public class Gato {
     @OneToOne(mappedBy = "gato")
     private Adopcion adopcionActual;   // opcional
 
-    public Gato(Long id, String nombre, String raza, Integer edad, String foto, EstadoSalud estadoDeSalud, String qr, boolean adoptado, Zona zona, HistorialMedico historialMedico, List<Tarea> tareas, Adopcion adopcionActual) {
+    public Gato(int id, String nombre, String raza, Integer edad, String foto, EstadoSalud estadoDeSalud, String qr, boolean adoptado, Zona zona, HistorialMedico historialMedico, List<Tarea> tareas, Adopcion adopcionActual) {
         this.id = id;
         this.nombre = nombre;
         this.raza = raza;
@@ -68,11 +68,11 @@ public class Gato {
 
     public Gato() {}
     
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
